Replace qt5-default with qtbase5-dev
Apt on Ubuntu now returns error saying the qt5-default package has no installation candidate. Instead we can use qtbase5-dev package as it is officially available on the repository. Also it is recommended from KDE community to not use the qt5-default package as a build dependency and to use qtbase5-dev instead. I also added the qt5-qmake package as the project uses qmake to generate makefiles.